add TaskFileData and support metrics for them
